# docs/uber-rag/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Uber-RAG API
  description: |
    API-first, ACL-aware RAG platform for textbooks and loose documents.
    Every UI action is reachable through this API. Backend enforces ACL at all layers.
  version: "0.1.0"
  contact:
    name: Uber-RAG Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /documents/{document_id}/acl:
    get:
      summary: Get ACL grants for a document
      operationId: getDocumentAcl
      tags: [ACL]
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: ACL grants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AclGrant"
    put:
      summary: Replace ACL grants for a document
      operationId: setDocumentAcl
      tags: [ACL]
      parameters:
        - name: document_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AclGrant"
      responses:
        "200":
          description: ACL updated
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC access token from Keycloak
  schemas:
...
    AclGrant:
      type: object
      properties:
        owner_user_id: { type: string, format: uuid }
        tenant_id: { type: string }
        allowed_user_ids:
          type: array
          items: { type: string, format: uuid }
        allowed_group_ids:
          type: array
          items: { type: string, format: uuid }
        visibility:
          type: string
          enum: [private, group, tenant, public]
        sensitivity:
          type: string
          enum: [public, internal, confidential, restricted]
        expires_at: { type: string, format: date-time, nullable: true }

    # ── Collections ────────────────────────────────────────────
...